requirements.txt
# HTTP requests and API communication
requests==2.31.0
urllib3==1.26.16

# RSS feed parsing
feedparser==6.0.10

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# OpenAI API for content generation
openai==1.3.8

# WordPress integration
python-wordpress-xmlrpc==2.3

# Database
sqlite3

# Task scheduling
schedule==1.2.0

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
jsonschema==4.19.2

# Logging
colorlog==6.7.0

# HTTP server (for testing)
flask==2.3.3

# Data validation
pydantic==2.4.2

# Rate limiting
ratelimit==2.2.1

# Configuration management
configparser==5.3.0

# Text processing
nltk==3.8.1
janome==0.4.2

# Image processing (for thumbnails)
Pillow==10.0.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0